+++ miniproject/src/main/java/com/iit/mp/dao/MemberDao.java
@@ -1,7 +1,7 @@
 package com.iit.mp.dao;
 
 import com.iit.mp.dto.MemberDto;
-import com.iit.mp.security.SecurityUserVO;
+import com.iit.mp.security.PrincipalDetails;
 
 public interface MemberDao {
 
@@ -12,7 +12,7 @@ public interface MemberDao {
 	/* 회원가입 */
 	void memberJoin(MemberDto memberDto);
 	/* 로그인 */
-//	MemberDto loginMember(String mbrId);
-	SecurityUserVO loginMember(String mbrId);
+	MemberDto loginMember(String mbrId);
+//	PrincipalDetails loginMember(String mbrId);
 	
 }
