+++ miniproject/src/main/java/com/iit/mp/dao/MemberDaoImpl.java
@@ -6,7 +6,7 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Repository;
 
 import com.iit.mp.dto.MemberDto;
-import com.iit.mp.security.SecurityUserVO;
+import com.iit.mp.security.PrincipalDetails;
 
 @Repository
 public class MemberDaoImpl implements MemberDao {
@@ -39,8 +39,8 @@ public class MemberDaoImpl implements MemberDao {
 	
 	/* 로그인 */
 	@Override
-//	public MemberDto loginMember(String mbrId) {
-	public SecurityUserVO loginMember(String mbrId) {
+	public MemberDto loginMember(String mbrId) {
+//	public PrincipalDetails loginMember(String mbrId) {
 		return sqlSession.selectOne("member.login", mbrId);
 	}
