--- miniproject/src/main/java/com/iit/mp/security/SecurityUserVO.java
@@ -1,63 +1,0 @@
-package com.iit.mp.security;
-
-import java.sql.Date;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import com.iit.mp.dto.MemberDto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data @AllArgsConstructor @NoArgsConstructor @Builder
-public class SecurityUserVO implements UserDetails {
-
-	private String mbrId, mbrPw, mbrNm, mbrMblTelno, mbrHomeAddr, mbrHomeDaddr, mbrGrd;
-	private Date mbrBrdt, mbrRegYmd;
-	
-	@Override
-	public Collection<? extends GrantedAuthority> getAuthorities() {
-		List<GrantedAuthority> auth = new ArrayList<GrantedAuthority>();
-		auth.add(new SimpleGrantedAuthority(getMbrGrd()));
-		return auth;
-	}
-	
-	@Override
-	public String getPassword() {
-		return getMbrPw();
-	}
-	
-	@Override
-	public String getUsername() {
-		return getMbrId();
-	}
-	
-	@Override
-	public boolean isAccountNonExpired() {
-		return true;
-	}
-	
-	@Override
-	public boolean isAccountNonLocked() {
-		return true;
-	}
-
-	@Override
-	public boolean isCredentialsNonExpired() {
-		return true;
-	}
-
-	@Override
-	public boolean isEnabled() {
-		return true;
-	}
-	
-}
